# docs/openapi/community-map-api.openapi.yaml
openapi: 3.1.0
info:
  title: Chengdu Tongzilin Community Map API
  version: 0.1.0
  description: |
    Community Map API for mobile, Mini Program, Admin, and Postman debugging.

    Current status:
    - Local Koa HTTP routes and shared contracts are implemented for the listed endpoints.
    - CloudBase dev URL is documented for deployment target only; `/api` route and live acceptance still require verification.
    - `x-mock-user-id` is a local/dev actor header and is not a production authentication scheme.
servers:
  - url: http://127.0.0.1:8787
    description: local Koa dev server
  - url: https://cloud1-d7gxdk8t43bd639c0.service.tcloudbase.com/api
    description: CloudBase dev HTTP access service, pending `/api` route verification
...
components:
  securitySchemes:
    MockActorHeader:
      type: apiKey
      in: header
      name: x-mock-user-id
      description: Local/dev mock actor id, for example `user_001` for admin flows.
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 10
    CommunityId:
      name: communityId
      in: query
      schema:
        type: string
        default: tongzilin
    Keyword:
      name: keyword
      in: query
      schema:
        type: string
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          examples:
            [
              VALIDATION_ERROR,
              UNAUTHORIZED,
              FORBIDDEN,
              NOT_FOUND,
              INTERNAL_ERROR
            ]
        message:
          type: string
        details: {}
    ApiFailure:
      type: object
      required: [success, error, requestId]
      properties:
        success:
          type: boolean
          const: false
        error:
          $ref: "#/components/schemas/ApiError"
        requestId:
          type: string
    SuccessEnvelope:
      type: object
      required: [success, data, requestId]
      properties:
        success:
          type: boolean
          const: true
        data: {}
        requestId:
          type: string
    Coordinates:
      type: object
      required: [latitude, longitude]
      properties:
        latitude:
          type: number
          example: 30.615
        longitude:
          type: number
          example: 104.062
...
    PlaceListItem:
      type: object
      required:
        [
          _id,
          name_zh,
          name_en,
          cover_url,
          category_level_1,
          category_level_2,
          short_address_zh,
          short_address_en,
          summary_zh,
          summary_en,
          tag_ids,
          is_recommended,
          recommended_reason_zh,
          recommended_reason_en,
          supports_navigation
        ]
      properties:
        _id:
          type: string
        name_zh:
          type: string
        name_en:
          type: string
        cover_url:
          type:
            - string
            - "null"
          format: uri
        category_level_1:
          type: string
        category_level_2:
          type: string
        short_address_zh:
          type: string
        short_address_en:
          type: string
        summary_zh:
          type: string
        summary_en:
          type: string
        tag_ids:
          type: array
          items:
            type: string
        is_recommended:
          type: boolean
        recommended_reason_zh:
          type:
            - string
            - "null"
        recommended_reason_en:
          type:
            - string
            - "null"
        supports_navigation:
          type: boolean
...
    PlaceGalleryMedia:
      type: object
      required: [file_id, cloud_path, url, alt_zh, alt_en]
      properties:
        file_id:
          type: string
        cloud_path:
          type: string
        url:
          type: string
          format: uri
        alt_zh:
          type: string
        alt_en:
          type: string
    PlaceDetail:
      type: object
      required:
        [
          _id,
          community_id,
          name_zh,
          name_en,
          cover_url,
          category_level_1,
          category_level_2,
          tag_ids,
          address_zh,
          address_en,
          location,
          business_hours_zh,
          business_hours_en,
          intro_zh,
          intro_en,
          gallery_media,
          gallery_urls,
          is_recommended,
          recommended_reason_zh,
          recommended_reason_en,
          supports_navigation,
          supports_favorite,
          supports_share,
          navigation,
          share
        ]
      properties:
        _id:
          type: string
        community_id:
          type: string
        name_zh:
          type: string
        name_en:
          type: string
        cover_url:
          type:
            - string
            - "null"
          format: uri
        category_level_1:
          type: string
        category_level_2:
          type: string
        tag_ids:
          type: array
          items:
            type: string
        address_zh:
          type: string
        address_en:
          type: string
        location:
          $ref: "#/components/schemas/Coordinates"
        business_hours_zh:
          type: string
        business_hours_en:
          type: string
        intro_zh:
          type: string
        intro_en:
          type: string
        gallery_media:
          type: array
          items:
            $ref: "#/components/schemas/PlaceGalleryMedia"
        gallery_urls:
          type: array
          items:
            type: string
            format: uri
        is_recommended:
          type: boolean
        recommended_reason_zh:
          type:
            - string
            - "null"
        recommended_reason_en:
          type:
            - string
            - "null"
        supports_navigation:
          type: boolean
        supports_favorite:
          type: boolean
        supports_share:
          type: boolean
        navigation:
          type: object
          required:
            [latitude, longitude, name_zh, name_en, address_zh, address_en]
          properties:
            latitude:
              type: number
            longitude:
              type: number
            name_zh:
              type: string
            name_en:
              type: string
            address_zh:
              type: string
            address_en:
              type: string
        share:
          type: object
          required: [title_zh, title_en, summary_zh, summary_en]
          properties:
            title_zh:
              type: string
            title_en:
              type: string
            summary_zh:
              type: string
            summary_en:
              type: string
    PlaceImportReviewMetadata:
      type:
        - object
        - "null"
      properties:
        source_type:
          type: string
          const: volunteer_spreadsheet
        source_file:
          type: string
        source_sheet:
          type: string
        source_column:
          type: string
        source_import_id:
          type: string
        volunteer_category_raw:
          type:
            - string
            - "null"
        category_code_candidate_raw:
          type:
            - string
            - "null"
        raw_fields:
          type: object
          additionalProperties:
            type: string
        review_blockers:
          type: array
          items:
            type: string
        review_notes:
          type: array
          items:
            type: string
    Place:
      allOf:
        - $ref: "#/components/schemas/PlaceDetail"
        - type: object
          required:
            [
              cover_file_id,
              tencent_map_poi_id,
              recommended_rank,
              gallery_file_ids,
              status
            ]
          properties:
            cover_file_id:
              type:
                - string
                - "null"
            tencent_map_poi_id:
              type:
                - string
                - "null"
            recommended_rank:
              type: integer
              minimum: 0
            gallery_file_ids:
              type: array
              items:
                type: string
            status:
              type: string
              enum: [draft, published, offline]
            import_review:
              $ref: "#/components/schemas/PlaceImportReviewMetadata"
    CreatePlaceInput:
      type: object
      required:
        [
          name_zh,
          name_en,
          category_level_1,
          category_level_2,
          address_zh,
          address_en,
          location,
          business_hours_zh,
          business_hours_en,
          intro_zh,
          intro_en
        ]
      properties:
        name_zh:
          type: string
        name_en:
          type: string
        cover_file_id:
          type:
            - string
            - "null"
          default: null
        cover_url:
          type:
            - string
            - "null"
          format: uri
          default: null
        category_level_1:
          type: string
        category_level_2:
          type: string
        tag_ids:
          type: array
          items:
            type: string
          default: []
        address_zh:
          type: string
        address_en:
          type: string
        location:
          $ref: "#/components/schemas/Coordinates"
        business_hours_zh:
          type: string
        business_hours_en:
          type: string
        intro_zh:
          type: string
        intro_en:
          type: string
        recommended_reason_zh:
          type:
            - string
            - "null"
          default: null
        recommended_reason_en:
          type:
            - string
            - "null"
          default: null
        is_recommended:
          type: boolean
          default: false
        recommended_rank:
          type: integer
          minimum: 0
          default: 0
        gallery_file_ids:
          type: array
          items:
            type: string
          default: []
        gallery_urls:
          type: array
          items:
            type: string
            format: uri
          default: []
        tencent_map_poi_id:
          type:
            - string
            - "null"
          default: null
        supports_navigation:
          type: boolean
          default: true
        supports_favorite:
          type: boolean
          default: true
        supports_share:
          type: boolean
          default: true
        status:
          type: string
          enum: [draft, published, offline]
          default: draft
        import_review:
          $ref: "#/components/schemas/PlaceImportReviewMetadata"
    UpdatePlaceInput:
      type: object
      description: Partial create place input. All fields from `CreatePlaceInput` are optional.
      properties:
        name_zh:
          type: string
        name_en:
          type: string
        cover_file_id:
          type:
            - string
            - "null"
        cover_url:
          type:
            - string
            - "null"
          format: uri
        category_level_1:
          type: string
        category_level_2:
          type: string
        tag_ids:
          type: array
          items:
            type: string
        address_zh:
          type: string
        address_en:
          type: string
        location:
          $ref: "#/components/schemas/Coordinates"
        business_hours_zh:
          type: string
        business_hours_en:
          type: string
        intro_zh:
          type: string
        intro_en:
          type: string
        recommended_reason_zh:
          type:
            - string
            - "null"
        recommended_reason_en:
          type:
            - string
            - "null"
        is_recommended:
          type: boolean
        recommended_rank:
          type: integer
          minimum: 0
        gallery_file_ids:
          type: array
          items:
            type: string
        gallery_urls:
          type: array
          items:
            type: string
            format: uri
        tencent_map_poi_id:
          type:
            - string
            - "null"
        supports_navigation:
          type: boolean
        supports_favorite:
          type: boolean
        supports_share:
          type: boolean
        status:
          type: string
          enum: [draft, published, offline]
        import_review:
          $ref: "#/components/schemas/PlaceImportReviewMetadata"
    DeletePlaceResponse:
      type: object
      required: [deleted_id]
      properties:
        deleted_id:
          type: string
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiFailure"
...
    Forbidden:
      description: Insufficient permission
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiFailure"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiFailure"
paths:
...
  /places:
    get:
      tags: [Places]
      operationId: listPlaces
      summary: List public places
      description: Public list v1 returns card fields only and never returns detail-only fields such as gallery media or navigation.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/CommunityId"
        - $ref: "#/components/parameters/Keyword"
        - name: category
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: recommended
          in: query
          schema:
            type: boolean
        - name: sort
          in: query
          schema:
            type: string
            enum: [recommended, name]
            default: recommended
      responses:
        "200":
          description: Place list page
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [items, page, pageSize, total]
                        properties:
                          items:
                            type: array
                            items:
                              $ref: "#/components/schemas/PlaceListItem"
                          page:
                            type: integer
                          pageSize:
                            type: integer
                          total:
                            type: integer
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /places/{id}:
    get:
      tags: [Places]
      operationId: getPlace
      summary: Get public place detail
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Place detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PlaceDetail"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/places:
    get:
      tags: [Admin, Places]
      operationId: adminListPlaces
      summary: List admin places
      security:
        - MockActorHeader: []
      responses:
        "200":
          description: Admin place page
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [items, page, pageSize, total]
                        properties:
                          items:
                            type: array
                            items:
                              $ref: "#/components/schemas/Place"
                          page:
                            type: integer
                          pageSize:
                            type: integer
                          total:
                            type: integer
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Admin, Places]
      operationId: adminCreatePlace
      summary: Create place
      security:
        - MockActorHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlaceInput"
      responses:
        "201":
          description: Created place
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Place"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/places/{id}:
    patch:
      tags: [Admin, Places]
      operationId: adminUpdatePlace
      summary: Update place
      description: Partial admin update. Omitted fields are preserved; invalid body values return the standard validation error envelope.
      security:
        - MockActorHeader: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlaceInput"
      responses:
        "200":
          description: Updated place
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Place"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Admin, Places]
      operationId: adminDeletePlace
      summary: Delete place
      description: Admin hard delete. The deleted place is removed from admin reads and from public list, marker, and detail reads.
      security:
        - MockActorHeader: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Deleted place id
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/DeletePlaceResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...