# packages/shared/src/types/entities.ts
import type { z } from "zod";

import type {
  AnnouncementSchema,
  AuthSessionSchema,
  CommentSchema,
  EventRegistrationSchema,
  EventSchema,
  EventTicketSchema,
  FileAssetSchema,
  NotificationSchema,
  PlaceSchema,
  PostSchema,
  UserSchema
} from "../schemas/entities";
import type {
  PlaceDetailSchema,
  DeletePlaceResponseSchema,
  PlaceGalleryMediaSchema,
  PlaceListItemSchema,
  PlaceMapMarkerSchema
} from "../schemas/places";

export type User = z.infer<typeof UserSchema>;
export type AuthSession = z.infer<typeof AuthSessionSchema>;
export type Event = z.infer<typeof EventSchema>;
export type EventRegistration = z.infer<typeof EventRegistrationSchema>;
export type EventTicket = z.infer<typeof EventTicketSchema>;
export type Place = z.infer<typeof PlaceSchema>;
export type PlaceListItem = z.infer<typeof PlaceListItemSchema>;
export type PlaceGalleryMedia = z.infer<typeof PlaceGalleryMediaSchema>;
export type PlaceDetail = z.infer<typeof PlaceDetailSchema>;
export type PlaceMapMarker = z.infer<typeof PlaceMapMarkerSchema>;
export type DeletePlaceResponse = z.infer<typeof DeletePlaceResponseSchema>;
export type Post = z.infer<typeof PostSchema>;
export type Comment = z.infer<typeof CommentSchema>;
export type Announcement = z.infer<typeof AnnouncementSchema>;
export type Notification = z.infer<typeof NotificationSchema>;
export type FileAsset = z.infer<typeof FileAssetSchema>;